Re-enabled audio
Moved the audio settings into their own component per the suggestion in #27.
