+++ src/components/BottomControls/index.tsx
@@ -1,13 +1,12 @@
-import React, { useContext } from "react";
+import React from "react";
 
-import { ConsentingAudioContext } from "../../hooks/useConsentingAudio";
 import { ButtonsList } from "../ButtonsList";
-import { InputButton } from "../Inputs/InputButton";
 import { InputLink } from "../Inputs/InputLink";
 import { Settings } from "../../types";
+import { BeepButton } from "../BeepButton";
+import { Notice } from "../Notice";
 import { ShareButton } from "../ShareButton";
 import { ZenZone } from "../ZenZone";
-import { Notice } from "../Notice";
 
 export type BottomControlsProps = {
     remaining: number;
@@ -18,8 +17,6 @@ export const BottomControls: React.FC<BottomControlsProps> = ({
     remaining,
     settings,
 }) => {
-    const { audio, toggleAudio } = useContext(ConsentingAudioContext);
-
     return (
         <ZenZone zen={settings.zen}>
             <Notice>
@@ -30,10 +27,7 @@ export const BottomControls: React.FC<BottomControlsProps> = ({
                     View source
                 </InputLink>
                 <ShareButton settings={settings} />
-                <InputButton
-                    onClick={toggleAudio}
-                    value={audio ? "Disable beeps" : "Enable beeps"}
-                />
+                <BeepButton over={-remaining} />
             </ButtonsList>
         </ZenZone>
     );
